componentize the form
